Object viewer class
This somewhat simple script allows the player to pan, zoom and rotate a particular object (in this case the banknotes) in order to easily view each part of it. This controls entirely using the mouse. Drag while holding right mouse to rotate, or the middle mouse to pan. Use the scroll wheel to zoom, and click the reset button on a HUD canvas to revert the position and rotation back to normal.

This code required me to set up an action asset using Unity's new Input System.