'Level' class for structuring game progression
Due to the relative simplicity of the game structure, each 'level' and its content can be determined programatically just by changing certain variables. By creating a ScriptableObject whose variables can be swapped in and out, I can easily generate new levels with different factors influencing the challenge, like the time limit, number of notes to check, the number of counterfeits and the kinds of defects that occur. The last factor isn't implemented into gameplay yet, but will be easy to do since the infrastructure is already set up.