Remade banknote prefab to work better as an object in physical space, rather than just being part of a canvas
The large size of the canvas required to have the desired level of detail meant I needed to drastically scale down its size to fit in the desired game space, especially when I decided to have the visible game environment be a 3D space. I changed the base object to have a regular transform rather than a RectTransform, and the canvas on which all the graphical elements of the note are oriented were added as a child, with an incredibly small scale. This way, I can have the banknote size be realistically small while still having a scale of one, preventing awkward scaling glitches when swapping between different parent transforms.