LevelManager for controlling level generation and win criteria
This MonoBehaviour class takes a reference to a ‘Level’ ScriptableObject and generates a series of notes with randomised defects for the player to inspect. It also controls additional code, such as:
•	Setting and pausing the game’s timer
•	Informing the player how many notes remain
•	Transitioning from one note to the next, with animations
•	Updating the object viewer to focus on the current note
•	Setting up the buttons that let the player judge if a note is real or counterfeit
•	Keeping track of the player’s judgements for each note (controlled with an array of bools)
Currently, nothing happens once the player processes every note, or if the timer runs out. But the code for switching to each new note works very well.