EndScreen class for wins or game overs
This class manages the menu that appears at the end of a level. It tells the player whether they succeeded or failed, as well as information about their performance such as incorrect guesses and remaining time, and will show different HUD elements based on the success/failure state and the reasons for said state.