ReferenceWindow class for showing defects that need to be watched out for
This class, when refreshed upon the generation of a new level, will use the defect flags in the current Level ScriptableObject, and generate a note with appropriate descriptions for each defect present in the current level. Each description will describe the way a particular part of the note should be, so the player can watch for counterfeit notes which don’t meet that criteria.