Changed button to show references, and reference text, to be more informative without requiring a tutorial
I changed the control from just being a regular on-screen button to resembling a piece of paper sticking up out of the bottom of the screen. This looks 'off' and catches the eye more easily. Additionally the wording on the note is changed to more clearly communicate that these are the correct qualities being shown, and that the player should check for notes that don't look like how they're described in the note.