Fixed bug where visible requirements did not represent actual requirements for completing a level
The fail state for incorrect judgements is determined by the number of errors you have to make in order to fail. But the end screen presents a 'minimum correct required to pass' value. Since this value was obtained just by subtracting the error number from the number of notes, you could meet the required number of guesses as specified in said screen but still be failed. I changed the code to simply add one to the visible amount, to be clearer and more informative.