AdvancedButton script
This script can be attached to an object the same way as a conventional 'Button' component, but includes extra functionality for when the player hovers over it. It doesn't have all the same functionality as the default Button class, but can be used alongside it to provide additional effects, such as cosmetic features that are distinct from regular gameplay ones.