MusicHandler for controlling dynamic music
This MonoBehaviour class exists in a level and contains a list of ‘MusicStateData’ structs. Each struct contains multiple tracks to randomly select, and a bool for if the music should loop. The kinds of music are differentiated using enums representing different states, e.g. ‘Active’, Menus’, ‘Win’, ‘Fail’ and so on. By running a public function (either using an enum as a parameter, or a string so it can be included in UnityEvents), I can easily switch between music tracks based on the current state of the game.