LevelByLevelMode class to decouple level-based gameplay mechanics from general gameplay controls
This script inherits from Gamemode, and includes all the functionality for loading a gameplay scenario based on a level. This functionality is already present in the class currently named LevelManager, but it will be removed from that class to decouple it from the code for gameplay controls, and allow other non-level-based gameplay modes to be implemented. I plan on adding an infinite mode, and may also implement a zen mode with no failure state.