Turned LevelManager into GameplayHandler, to decouple level code from general gameplay controls
I discovered that implementing new gameplay modes would be much harder than I anticipated, as many mechanics related to level transitions were hard-coupled to the general gameplay controls, and other code that would need to be re-used regardless of the mode. To this extent, I overhauled LevelManager, by removing all level-based code (and placing it in LevelByLevelMode), and renaming it to GameplayHandler so as to be more informative. Now, whatever Gamemode class is loaded will then reference the persistent GameplayHandler in order to set up the gameplay based on whatever gamemode and level is being requested.