Change to AdvancedSelectable - buttons with visible names have their text assigned automatically based on the object's name
This makes it easier in the editor to rename a button to have its usage be clear to both the player and the developer. This setting can be disabled by unchecking 'autoSetVisibleName', a value that is set to private so it is not confused as being part of the class' actual function.