Removed 'NextItem' function from Gamemode class
This function was only used in one single place in a function, so I took the code and put it there. I was then able to remove it entirely for a simpler class with less variables to update, as it was previously an abstract function that needed to be overridden in inherited classes.