Decoupled timer code from GameplayHandler and coupled it with LevelByLevelMode
The current timer code is designed exclusively to count down, which is a problem considering I want the timer in the infinite mode to count up. Additionally, I may include a zen mode which has no timer at all. To that end I've removed the timer code from GameplayHandler and put it in LevelByLevelMode.