Fixed bug with trying to reference destroyed banknotes upon switching to gameplay
These bugs would occur when the player quit out of gameplay to the main menu, then entered a gameplay mode again. This problem existed due to dodgy code for refreshing the play space.
I fixed this with an abstract PurgeItems() function in the Gamemode class. This has different variants for the level-based and infinite modes, since the former has multiple notes active at a time, and the infinite mode needs only one.
I also took advantage of the change to fix some redundancy. I realised the ‘allItems’ data in Gamemode was only necessary for LevelByLevelMode, since every generic reference to the notes in the scene was taken care of with abstract functions/properties, so I moved that variable to that class instead. This made the spawning and purging code in InfiniteMode much cleaner and more intuitive.