Fixed bug where items weren't spawning in from the appropriate locations
This is a cosmetic one, but one easily fixed with some lines of code in GameplayHandler to move each new note to the starting position when they are ready to be shown. I also removed some code from LevelByLevelMode that did the same thing, since the general purpose code in GameplayHandler made it redundant.