LifeMeterGenerator class for setting up large 'heart'-type health meters
This script is used to automatically generate a grid of hearts, for large health meters. It's not strictly necessary for this use case (since I realised with the small number of lives I intend it isn't hard to add each heart manually), but I implemented it anyway in case I need to use it for other reasons.