Removed 'ka-ching!' noise when marking an object as correct
I realised this could provide inaccurate feedback by making the player think they've made a correct judgement, if marking a fake note as real. Plus, it means I can re-use it for the infinite mode, when the player does need to be immediately informed of if they were correct or not.