Fixed bug where JudgementHandler's OnResetGame() function is never run.
I realised there was no code to run this. To fix this, I made an abstract array in Gamemode to return a list of appropriate modifiers, which is then overridden in inherited versions with whatever modes are necessary. I then created a simple for loop in EnterGame(), that activates each modifier.