Incremental values in VolumeSlider
I realised that making the min and max slider values the minimum and maximum required for the logarithmic conversion to work was causing a problem. By making it a float with a smooth transition, it could make getting the precise desired volume annoying, and I was using a clamp function I didn’t need to limit it like that anyway. I decided to change this class to allow for simplified and more intuitive UI navigation, as that level of fine control isn’t necessary.
I realised that the Slider class has a ‘normalizedValue’ property, that allows the set value to be obtained or updated as if it was a 0-1 float, rather than the value specified by the slider’s range. This means I can make the slider values whatever I want for ease of use, while getting the correct value in code.