ObjectViewer: renamed OnReset() to ResetViewOrientation() for improved clarity
It being named OnReset() was a holdover from when the function was activated through the PlayerInput class and InputActionAsset system. Now that it's enabled through a button, I renamed it to communicate that it's the function that resets the view, not an incidental function that happens once the view is reset.