Replaced Timer code with Timekeeper code
I retained the Timer name since it's easier to type. The text for the time display is located in GameplayHandler since I intend to move all the timer references to that class, so only one reference needs to be assigned.

I also added two extra properties - 'elapsed' and 'remaining', which return the elapsed and remaining seconds converted into easy TimeValue structs. This was a function present in the old Timer class, that I preserved to allow for easy display of values in the HUD and console.

I also moved the TimeValue struct to continue being inside the Timer class, since the Level class depended on it and changing that value would mean re-assigning all the time values for every level.