Added base 'currentItemIndex' property in 'Gamemode'
I realised that InfiniteMode and LevelByLevelMode both had separate variables that served to track the number of notes the player had checked. I combined them into one variable included in the 'Gamemode' class.

I also added the code to increment the number in a base version of the PrepareNextItem() function, so I didn't need to write it again in InfiniteMode. I did keep it in LevelByLevelMode, since it completely overrides the function and won't increment the number if there aren't any more to check.