Simplified random note generation code in InfiniteMode
The code was originally supposed to be more complex and use noise to ensure there weren't too many long streaks of notes being real or fake, but I couldn't get it to work properly. I also realised that having those streaks might be beneficial to gameplay, as the player must avoid being complacent from a streak of one kind of items, and make an incorrect judgement when the next item is different.