Added code in Gamemode to reset current item index
Again, this function occurred twice in both InfiniteMode and LevelByLevelMode, so I decided to simplify it.
I also realised that due to the way the incrementing code is called, in InfiniteMode the index would be one higher than it's supposed to be, because the index is incremented before the first note is spawned. In LevelByLevelMode this wasn't a problem because the index was reset to -1 instead of 0, so I made sure that this was the case in Gamemode.