Bugfix - TimeValue struct does not display values correctly from negative seconds float
I realised this was due to a glitch in how the values were generated, and that there wasn't a clear way to indicate a negative value. I decided to instead have a single 'negative' bool. In the constructor that uses a single float for total seconds, the bool is assigned based on if the float is less than zero. The hour, minute and second values are assigned using an absolute version of the initial float parameter, to ensure it's all correctly assigned. The ToString() function now has a simple bool adding a minus value if the time struct is negative.