Changed GameplayHandler gamemode assignment from a property to a function
This change both simplifies the code and makes the procedure more intuitive. By simply changing a variable a developer might not realise they are running a significant amount of code, and someone who does want to generate the game may not realise it because there isn't a specific function for it.