(UNTESTED) SaveHandler class for saving and loading data
This static class will save game data to and load it from a JSON file. Saving and loading are performed effortlessly using Save() and Load() functions, which will assign the data to an appropriate file based on an index.
Save data is stored during runtime in custom 'SaveFile' structs. A public static instance of this struct, named 'currentData', can be read from and written to by any other game element, when important data needs to be updated. Two static System.Actions, onLoadGame and onSaveGame, allow any number of actions to occur on saving or loading, by adding them as listeners.
‘currentData’, ‘onLoadGame’ and ‘onSaveGame’ allow the rest of the code to effortlessly read from or write to save data, and act on the player saving or loading, without the system itself needing to reference any other class or scene object.

(The SaveFile class presently has no variables included in this commit, because they aren't relevant to the way the save system works)